+++ CHANGELOG.md
@@ -20,7 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `CreatesOrUpdatesUser` trait
 - Added newly added permissions in `ApiSeed` command
 - Added `user.unverified.new_only` config value
-- Added user impersonation
+- Added user impersonation and related event
 
 ### Changed